vk: rt: move metapass stuff into meatpipe
Metapass was a mistake -- moving resource acquisition and dispatch back
into meatpipe allows meatpipe to become fully opaque, and also more
tightly control its state wrt external resources and what not.
